multiple
Added references to Exceptions, in order to be able to see why the
operation failed, if it does, after the fact, by looking at sex, wex,
ex.

Crypto Server IP address setting is now done at program start.

Refactored the Crypto blocks, to reduce the amount of code that has to
be retyped for every crypto operation that one might do.